+++ PERCS/FT/ft_natives.h
@@ -1,6 +1,7 @@
 #include <stdint.h>
 
 extern "C" {
-void execute_plan(int64_t, double*, double*, signed int, signed int, signed int);
-int64_t create_plan(signed int, signed int, signed int);
+void execute_plan(int64_t, double*, double*, int64_t, int64_t, int64_t);
+int64_t create_plan(int64_t, int32_t, int32_t);
+
 }
